Restructure folders as a package
