+++ slate-editor/SlateEditor.jsx
@@ -2,24 +2,26 @@ import React from "react";
 import { Editor } from "slate-react";
 import { Value } from "slate";
 import PropTypes from "prop-types";
-import axios from "axios";
 
-import { initialValue, Toolbar } from "./slate-editor-core";
+import { initialValue, Toolbar } from "./core";
 
 // Inline
-import renderInline from "./slate-editor-inline";
+import renderInline from "./inline";
 
 // Link
-import { onPasteLink, renderLinkDialogWindow } from "./slate-editor-inline/link";
+import { onPasteLink, renderLinkDialogWindow } from "./inline/link";
 
 // Mark
-import { onKeyDown, renderMark } from "./slate-editor-marks";
+import { onKeyDown, renderMark } from "./marks";
 
 // Block
-import { renderBlock } from "./slate-editor-blocks";
+import { renderBlock } from "./blocks";
 
 // POST and GET url's
 import { POST_URL, GET_URL } from "../env";
+
+// DB interactions
+import { save, get, put, remove } from "../db";
 
 class SlateEditor extends React.Component {
 	// Constructor
@@ -113,75 +115,6 @@ class SlateEditor extends React.Component {
 		return { __html: this.editor.el.innerHTML };
 	};
 
-	// DB interactions
-	saveToDB = () => {
-		axios.post(POST_URL, {
-			content: this.state.value,
-		}).then((res) => {
-			localStorage.setItem("CONTENT_ID", res.data);
-			alert("Successfully saved to DB!");
-		}).catch(error => {
-			console.log(error);
-		});
-	}
-
-	loadFromDB = () => {
-		const _id = localStorage.getItem("CONTENT_ID");
-		if (!_id) {
-			alert("You have to save something first.");
-		} else {
-			axios.get(GET_URL, {
-				params: {
-					_id,
-				},
-			}).then(res => {
-				const content = res.data.content;
-				localStorage.setItem("CONTENT", JSON.stringify(content));
-				this.setState({
-					value: Value.fromJSON(content),
-				});
-				alert("Successfully loaded from DB!");
-			}).catch(error => {
-				console.log(error);
-			});
-		}
-	}
-
-	updateToDB = () => {
-		const _id = localStorage.getItem("CONTENT_ID");
-		if (!_id) {
-			alert("You have to save something first.");
-		} else {
-			axios.put(POST_URL, {
-				_id,
-				content: this.state.value,
-			}).then((res) => {
-				alert("Successfully updated to DB!");
-			}).catch(error => {
-				console.log(error);
-			});
-		}
-	}
-
-	deleteFromDB = () => {
-		const _id = localStorage.getItem("CONTENT_ID");
-		if (!_id) {
-			alert("You have to save something first.");
-		} else {
-			axios.delete(POST_URL, {
-				params: {
-					_id,
-				},
-			}).then(() => {
-				localStorage.removeItem("CONTENT");
-				localStorage.removeItem("CONTENT_ID");
-				alert("Successfully deleted from DB!");
-			}).catch(error => {
-				console.log(error);
-			});
-		}
-	}
-
 	render() {
 
 		// Global Context Destructuring
@@ -234,10 +167,10 @@ class SlateEditor extends React.Component {
 					/>
 					{isDialog && renderLinkDialogWindow(this.setWrapperRef, editor, value, cursorPosition)}
 				</div>
-				<button onClick={this.saveToDB}>Save to DB</button>
-				<button onClick={this.loadFromDB}>Load from DB</button>
-				<button onClick={this.updateToDB}>Update to DB</button>
-				<button onClick={this.deleteFromDB}>Delete from DB</button>
+				<button onClick={() => save(POST_URL, value)}>Save to DB</button>
+				<button onClick={() => get(this, GET_URL)}>Load from DB</button>
+				<button onClick={() => put(POST_URL, value)}>Update to DB</button>
+				<button onClick={() => remove(POST_URL)}>Delete from DB</button>
 				<div>
 					<p>State (JSON object):</p>
 					<pre
@@ -283,6 +216,7 @@ SlateEditor.defaultProps = {
 	ul: false,
 	textAlign: null,
 	link: false,
+	formats: null,
 };
 
 SlateEditor.propTypes = {
@@ -307,6 +241,7 @@ SlateEditor.propTypes = {
 		PropTypes.string,
 	]),
 	link: PropTypes.bool,
+	formats: PropTypes.array,
 };
 
 export default SlateEditor;
