+++ slate-editor/blocks/hasBlock.js
