+++ slate-editor/core/index.js
