+++ slate-editor/core/initialValue.js
