+++ slate-editor/formats/text-align/hasAlign.js
