+++ slate-editor/formats/text-align/hasMultipleAligns.js
