+++ slate-editor/formats/text-align/index.js
