+++ slate-editor/formats/text-align/onClickAlign.js
