+++ slate-editor/marks/hasMark.js
