+++ slate-editor/marks/renderMarkButton.jsx
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 
-import { Button } from "../slate-editor-core";
+import { Button } from "../core";
 
 import hasMark from "./hasMark";
 import onClickMark from "./onClickMark";
